refactor(dht): move file related stuff into own mod
